cosmosdb - move tests to cosmos_test package
