+++ azurerm/internal/services/cosmos/cosmosdb_account_data_source_test.go
@@ -1,4 +1,4 @@
-package tests
+package cosmos_test
 
 import (
 	"fmt"
