+++ azurerm/internal/services/cosmos/cosmosdb_account_resource_failover_test.go
@@ -1,4 +1,4 @@
-package tests
+package cosmos_test
 
 import (
 	"fmt"
