+++ azurerm/internal/services/cosmos/cosmosdb_gremlin_database_resource_test.go
@@ -1,4 +1,4 @@
-package tests
+package cosmos_test
 
 import (
 	"fmt"
